+++ engine/include/renderer/polygon_fill_mode.h
@@ -3,7 +3,7 @@
 #include "vlkypch.h"
 
 namespace Vulkyrie::Renderer {
-    enum class PolygonFillMode : u8 {
+    enum class PolygonFillMode : u32 {
         Fill,
         Line,
         Point,
